@@ -22,6 +22,7 @@
     "@supabase/auth-helpers-nextjs": "^0.5.2",
     "@supabase/auth-helpers-react": "^0.3.1",
     "@supabase/auth-ui-react": "^0.2.6",
+    "@supabase/ssr": "^0.1.0",
     "@supabase/supabase-js": "^2.0.5",
     "@testing-library/user-event": "^14.4.3",
     "@types/seedrandom": "^3.0.6",
@@ -33,11 +34,11 @@
     "date-fns": "^2.29.3",
     "framer-motion": "^9.1.6",
     "gray-matter": "^4.0.3",
-    "next": "12.3.1",
+    "next": "^14.1.3",
     "nodemailer": "^6.9.3",
-    "react": "18.2.0",
+    "react": "^18.2.0",
     "react-chartjs-2": "^5.1.0",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.39.3",
     "react-markdown": "^8.0.5",
     "remark-gfm": "^3.0.1",
@@ -57,12 +58,12 @@
     "@testing-library/react": "^13.4.0",
     "@types/node": "18.11.3",
     "@types/nodemailer": "^6.4.8",
-    "@types/react": "18.0.21",
-    "@types/react-dom": "18.0.6",
+    "@types/react": "18.2.48",
+    "@types/react-dom": "18.2.18",
     "@vanilla-extract/next-plugin": "^2.1.1",
     "autoprefixer": "^10.4.16",
     "eslint": "8.26.0",
-    "eslint-config-next": "12.3.1",
+    "eslint-config-next": "^14.1.3",
     "eslint-plugin-storybook": "^0.6.12",
     "jest": "^29.3.1",
     "jest-environment-jsdom": "^29.3.1",
