@@ -1,4 +1,4 @@
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
 export const useLocalStorageAuthFromParams = () => {
   const router = useRouter();
