@@ -1,11 +1,15 @@
+"use client";
 import React, { ReactElement } from "react";
 import Image from "next/image";
-import { useRouter } from "next/router";
+import { useRouter, usePathname } from "next/navigation";
 import { useAuthModal } from "components/context/EmailAuthModal";
-import { useUserSession } from "components/context/UserSessionContext";
+import { signout } from "@/actions/actions";
 
-export const SignupButton = (): ReactElement => {
-  const { user, signOut } = useUserSession();
+interface Props {
+  userId?: string;
+}
+
+export const SignupButton = ({ userId }: Props): ReactElement => {
   const router = useRouter();
 
   const { setIsOpen } = useAuthModal();
@@ -14,37 +18,38 @@ export const SignupButton = (): ReactElement => {
     router.push("/profile");
   };
 
+  const pathname = usePathname();
+
   // TODO: I don't love adding "if x route, do y" logic for generic components like this, since it degrades the reusability of the component
   // when we do a UI refactor, we should allow each route to render its own custom header component, and this won't be needed anymore.
-  const isUserProfileRoute = router.pathname === "/profile";
+  const isUserProfileRoute = pathname === "/profile";
 
   return (
     <>
-      {user ? (
+      {userId ? (
         isUserProfileRoute ? (
-          <button
-            className=" h-10 py-2 px-4 border-2 font-bold text-white border-white bg-transparent flex items-center rounded-md
+          //@ts-ignore
+          <form action={signout}>
+            <button
+              className=" h-10 py-2 px-4 border-2 font-bold text-white border-white bg-transparent flex items-center rounded-md
        hover:bg-white hover:text-black hover:shadow-NavShadow hover:cursor-pointer"
-            onClick={signOut}
-          >
-            Sign Out
-          </button>
+            >
+              Sign Out
+            </button>
+          </form>
         ) : (
           <button onClick={onProfile}>
             <Image
               src="/profile-icon.png"
               alt="profile icon"
-              width="60px"
-              height="50px"
+              width={60}
+              height={50}
               className="hover:shadow-NavShadow hover:cursor-pointer"
             />
           </button>
         )
       ) : (
-        <button
-          className="btn-main"
-          onClick={setIsOpen}
-        >
+        <button className="btn-main" onClick={setIsOpen}>
           Sign up / Log In!
         </button>
       )}
