@@ -1,20 +1,20 @@
-import { useUserSession } from "components/context/UserSessionContext";
 import React from "react";
 import { EmailAuthModal } from "../EmailAuthModal";
-import { Loading } from "../Loading";
 
-export const SignInGate = ({ children }: { children: React.ReactElement }) => {
-  const { isLoading, user } = useUserSession();
-
-  if (isLoading) {
-    return <Loading />;
-  }
-
-  return !!user ? (
+export const SignInGate = ({
+  children,
+  userId,
+  redirectUrl,
+}: {
+  children: React.ReactElement;
+  userId?: string;
+  redirectUrl: string;
+}) => {
+  return !!userId ? (
     <>{children}</>
   ) : (
     <EmailAuthModal
-      redirectUrl={window.location.href}
+      redirectUrl={redirectUrl}
       isOpen={true}
       titleOverride="Get your Login link"
     />
