@@ -1,3 +1,4 @@
+"use client";
 import {
   Toast,
   ToastClose,
@@ -5,11 +6,11 @@ import {
   ToastProvider,
   ToastTitle,
   ToastViewport,
-} from "components/ui/toast"
-import { useToast } from "components/ui/use-toast"
+} from "components/ui/toast";
+import { useToast } from "components/ui/use-toast";
 
 export function Toaster() {
-  const { toasts } = useToast()
+  const { toasts } = useToast();
 
   return (
     <ToastProvider>
@@ -25,9 +26,9 @@ export function Toaster() {
             {action}
             <ToastClose />
           </Toast>
-        )
+        );
       })}
       <ToastViewport />
     </ToastProvider>
-  )
+  );
 }
